refactor(chat-input): move reasoning effort slider into model picker sheet footer
The reasoning control no longer lives as a floating panel toggled from the
chat-input toolbar. It now renders pinned below the model list in the model
picker sheet (ModelPickerBottomSheet `footer` slot), separated by a divider,
since the selected model is what decides the available stops.

- Add ChatInputReasoningSection (divider + safe-area padding self-contained,
  renders nothing for models without reasoning stops) and
  useChatInputReasoningEffortSync (validity sync now lives with the composer,
  not the sheet)
- Delete ChatInputReasoningPanel, the toolbar pill/meter in ChatInputSurface,
  isReasoningPanelOpen state, and the effort-driven composer expansion
- Drop the slider's FadeIn/FadeOut entering animations (first-frame
  full-opacity flash) and pre-light the pixel field when opening at the
  trigger stop
- Sweep now-dead reasoning helpers (cycle/meter/off/active/tag) and their
  tests
- e2e: flow 00 opens the picker via chat-input-model-button and closes it by
  tapping the scrim (a downward swipe now lands on the list and scrolls)
