+++ src/screens/ChatScreen/NewTopicScreen.tsx
@@ -7,8 +7,7 @@ import { useFloatingChatInputLayout } from './workspace/hooks/useFloatingChatInp
 
 export function NewTopicScreen() {
   const { t } = useTranslation();
-  const { contentBottomInset, handleInputHeightChange, inputHeightShared } =
-    useFloatingChatInputLayout();
+  const { contentBottomInset, handleInputHeightChange } = useFloatingChatInputLayout();
 
   return (
     <ChatWorkspaceFrame>
@@ -26,7 +25,7 @@ export function NewTopicScreen() {
           {t('chat.newTopic.description')}
         </Text>
       </View>
-      <ChatComposer inputHeight={inputHeightShared} onHeightChange={handleInputHeightChange} />
+      <ChatComposer onHeightChange={handleInputHeightChange} />
     </ChatWorkspaceFrame>
   );
 }
