+++ src/screens/ChatScreen/input/ChatInput.tsx
@@ -11,10 +11,12 @@ import {
 import { isUniqueModelId } from '@/data/types/model';
 import { useModelById, useProviders, useTopic } from '@/hooks/chat';
 import { ChatInputActionSheet } from '@/screens/ChatScreen/input/components/ChatInputActionSheet';
+import { ChatInputReasoningSection } from '@/screens/ChatScreen/input/components/ChatInputReasoningSection';
 import {
   type ChatInputSendPayload,
   ChatInputSurface,
 } from '@/screens/ChatScreen/input/components/ChatInputSurface';
+import { useChatInputReasoningEffortSync } from '@/screens/ChatScreen/input/hooks/useChatInputReasoningEffortSync';
 import { useChatInputReasoningEfforts } from '@/screens/ChatScreen/input/hooks/useChatInputReasoningEfforts';
 import { createChatInputMessageParts } from '@/screens/ChatScreen/input/utils/chatInputAttachments';
 import { useChatRuntimeTopic } from '@/screens/ChatScreen/runtime';
@@ -46,6 +48,7 @@ export function ChatInput({ topicId }: ChatInputProps) {
       )
     : undefined;
   const reasoningEfforts = useChatInputReasoningEfforts();
+  useChatInputReasoningEffortSync(reasoningEfforts);
 
   const openModelPicker = useCallback(() => {
     modelPickerRef.current?.present();
@@ -84,10 +87,10 @@ export function ChatInput({ topicId }: ChatInputProps) {
         onModelPickerPress={openModelPicker}
         onSendPress={handleSendPress}
         onStopPress={chatRuntime.abort}
-        reasoningEfforts={reasoningEfforts}
       />
       <ChatInputActionSheet />
       <ModelPickerBottomSheet
+        footer={<ChatInputReasoningSection />}
         onSelect={handleModelSelect}
         ref={modelPickerRef}
         selectedModelId={selectedModelId}
