+++ src/screens/ChatScreen/workspace/components/ChatComposer.tsx
@@ -1,27 +1,21 @@
-import type { SharedValue } from 'react-native-reanimated';
-
-import { ChatInputReasoningPanel } from '../../input/components/ChatInputReasoningPanel';
 import { ChatInputProvider } from '../../input/context/ChatInputProvider';
 import { FloatingChatInput } from './FloatingChatInput';
 
 type ChatComposerProps = {
   onHeightChange: (height: number) => void;
-  /** Live floating-input height; positions the reasoning panel above it. */
-  inputHeight: SharedValue<number>;
   topicId?: string;
 };
 
 /**
- * The floating composer plus its reasoning panel, wrapped in the shared
- * ChatInputProvider. Keeping the three together in one component means every
- * screen that shows the input (chat + new-topic) gets the provider and the
- * panel, so neither can drift out of the tree again.
+ * The floating composer wrapped in the shared ChatInputProvider, so every
+ * screen that shows the input (chat + new-topic) gets the provider with it.
+ * The reasoning-effort control lives inside the model picker sheet
+ * (ChatInputReasoningSection), not as a separate floating panel.
  */
-export function ChatComposer({ onHeightChange, inputHeight, topicId }: ChatComposerProps) {
+export function ChatComposer({ onHeightChange, topicId }: ChatComposerProps) {
   return (
     <ChatInputProvider>
       <FloatingChatInput onHeightChange={onHeightChange} topicId={topicId} />
-      <ChatInputReasoningPanel inputHeight={inputHeight} />
     </ChatInputProvider>
   );
 }
